requirements.txt
astroid==2.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Deprecated==1.2.6
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
PyGithub==1.43.8
PyJWT==1.7.1
pylint==2.4.1
requests==2.22.0
schedule==0.6.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
